Refactored Health Api methods in SvixClient to Health.cs
