+++ csharp/Svix/Abstractions/ISvixClient.cs
@@ -5,12 +5,6 @@ namespace Svix.Abstractions
 {
     public interface ISvixClient
     {
-        #region " Health "
-        
-        bool IsHealthy(string idempotencyKey = default);
-
-        Task<bool> IsHealthyAsync(string idempotencyKey = default, CancellationToken cancellationToken = default);
-        
-        #endregion
+        public Health Health { get; }
     }
 }
